requirements.txt
asgiref==3.5.2
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
distlib==0.3.6
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dulwich==0.20.50
filelock==3.8.2
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.11.0
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.9.2
platformdirs==2.6.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
psycopg2==2.9.5
ptyprocess==0.7.0
PyJWT==2.6.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.7
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.17.1
webencodings==0.5.1
zipp==3.11.0
